phoronix-c-ray-docker/requirements.txt
py-cpuinfo~=8.0.0
